Move exposed libs to pkg
