+++ pkg/kafka/kafka_offset_parser.go
