+++ pkg/kafka/kafka_offset_parser_test.go
