+++ pkg/kafka/log.go
