+++ pkg/kafka/mocks/mock_kafka_reader.go
@@ -5,7 +5,7 @@ package mocks
 import (
 	"context"
 
-	"github.com/xataio/pgstream/internal/kafka"
+	"github.com/xataio/pgstream/pkg/kafka"
 )
 
 type Reader struct {
